third_party/blink/web_tests/svg/custom/keyTimes-parsing-error.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests parsing of keyTimes attribute.");

var animate = document.createElementNS("http://www.w3.org/2000/svg", "animate");
animate.setAttribute("keyTimes", ";;");
animate.setAttribute("keyTimes", "0;.25;.5;1;;");

var successfullyParsed = true;
</script>
</body>
</html>
