third_party/blink/web_tests/svg/custom/keyTimes-parsing-error-expected.txt

CONSOLE ERROR: Error: <animate> attribute keyTimes: Invalid value, ";;".
CONSOLE ERROR: Error: <animate> attribute keyTimes: Invalid value, "0;.25;.5;1;;".
Tests parsing of keyTimes attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
